requirements.txt
numpy==1.26.4
pandas==1.5.3
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
crewai==0.95.0
crewai-tools==0.25.8